Added conditional checks before purging
Checking for valid cstate, query results count before initiating purging.

+++ bin/purge_user_timeseries.py
@@ -13,13 +13,15 @@ from bson import ObjectId
 import json
 from uuid import UUID
 import tempfile
+from datetime import datetime
 
 DEFAULT_DIR_NAME = tempfile.gettempdir()
 DEFAULT_FILE_PREFIX = "old_timeseries_"
 
 def exportOldTimeseriesAsCsv(user_id, all_data, filename):
     logging.info("Exporting data to CSV...")
-    filename += ".csv"
+    current_datetime = datetime.now().strftime("%Y%m%d_%H%M%S_%f")
+    filename += "_" + current_datetime + ".csv"
     all_df = pd.json_normalize(all_data)
     all_df.to_csv(filename)
     logging.info("Old timeseries data exported as CSV to {}".format(filename))
@@ -31,9 +33,10 @@ def exportOldTimeseriesAsJson(user_id, all_data, filename):
             return str(obj)
         raise TypeError(f"Type {type(obj)} not serializable")
 
-    filename += ".json"
+    current_datetime = datetime.now().strftime("%Y%m%d_%H%M%S_%f")
+    filename += "_" + current_datetime + ".json"
     with open(filename, 'w') as file:
-        json.dump(all_data, file, default=custom_encoder)
+        json.dump(all_data, file, default=custom_encoder, indent=4)
     logging.info("Old timeseries data exported as JSON to {}".format(filename))
 
 exportOptions = {
@@ -48,29 +51,38 @@ def purgeUserTimeseries(exportFileFlags, user_uuid, user_email=None, dir_name=DE
         user_id = ecwu.User.fromEmail(user_email).uuid
 
     cstate = esp.get_current_state(user_id, ecwp.PipelineStages.CREATE_CONFIRMED_OBJECTS)
-    last_ts_run = cstate['last_ts_run']
-    logging.info(f"last_ts_run : {last_ts_run}")
+    
+    if cstate is None:
+        logging.info(f"No matching pipeline state found for {user_id}, purging aborted.")
+    else:
+        last_ts_run = cstate['last_ts_run']
+        logging.info(f"last_ts_run : {last_ts_run}")
 
-    if not last_ts_run:
-        logging.warning("No processed timeseries for user {}".format(user_id))
-        exit(1)
+        if not last_ts_run:
+            logging.warning("No processed timeseries for user {}, purging aborted".format(user_id))
+            exit(1)
 
-    filename = dir_name + "/" + file_prefix + str(user_id)
-    logging.info("Querying data...")
-    all_data = list(edb.get_timeseries_db().find({"user_id": user_id, "metadata.write_ts": { "$lt": last_ts_run}}))
-    logging.info("Fetched data...")
+        filename = dir_name + "/" + file_prefix + str(user_id)
+        logging.info("Querying data...")
+        all_data = list(edb.get_timeseries_db().find({"user_id": user_id, "metadata.write_ts": { "$lt": last_ts_run}}))
 
-    if unsafe_ignore_save is True:
-        logging.warning("CSV export was ignored")
-    else: 
-        for key in exportFileFlags:
-            logging.info(f"{key} = {exportFileFlags[key]}")
-            if exportFileFlags[key] is True:
-                exportOptions[key](user_id, all_data, filename)
+        print(len(all_data))
 
-    logging.info("Deleting entries from database...")
-    result = edb.get_timeseries_db().delete_many({"user_id": user_id, "metadata.write_ts": { "$lt": last_ts_run}})
-    logging.info("{} deleted entries since {}".format(result.deleted_count, datetime.fromtimestamp(last_ts_run)))
+        if len(all_data) == 0:
+            logging.info("No matching data found for the user, purging aborted")
+        else:
+            logging.info("Fetched data, starting export")
+            if unsafe_ignore_save is True:
+                logging.warning("CSV export was ignored")
+            else: 
+                for key in exportFileFlags:
+                    logging.info(f"{key} = {exportFileFlags[key]}")
+                    if exportFileFlags[key] is True:
+                        exportOptions[key](user_id, all_data, filename)
+
+            logging.info("Deleting entries from database...")
+            result = edb.get_timeseries_db().delete_many({"user_id": user_id, "metadata.write_ts": { "$lt": last_ts_run}})
+            logging.info("{} deleted entries since {}".format(result.deleted_count, datetime.fromtimestamp(last_ts_run)))
     
 if __name__ == '__main__':
     logging.basicConfig(level=logging.DEBUG)
